# ux/documentation/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/llm/suggest_term_definition/stream": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Suggest Term Definition Stream",
        "description": "Stream term definition suggestion using specified flavor with Server-Side Events",
        "operationId": "suggest_term_definition_stream_api_llm_suggest_term_definition_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefinitionSuggestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/llm/suggest_layer_definition/stream": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Suggest Layer Definition Stream",
        "description": "Stream layer definition suggestion using specified flavor with Server-Side Events",
        "operationId": "suggest_layer_definition_stream_api_llm_suggest_layer_definition_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayerDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/llm/suggest_domain_definition/stream": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Suggest Domain Definition Stream",
        "description": "Stream domain definition suggestion using specified flavor with Server-Side Events",
        "operationId": "suggest_domain_definition_stream_api_llm_suggest_domain_definition_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/llm/suggest_term_definition": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Suggest Term Definition",
        "description": "Generate a term definition suggestion based on provided context using LLM.\n\nThis endpoint uses a Langchain pipeline with OpenAI GPT models to generate\ncontextually-aware term definitions based on domain context, hierarchical\nrelationships, component terms, and reference source information.\nSupports optional flavor parameter to use different LLM configurations.",
        "operationId": "suggest_term_definition_api_llm_suggest_term_definition_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefinitionSuggestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/llm/suggest_layer_definition": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Suggest Layer Definition",
        "description": "Generate a layer definition suggestion based on provided context using LLM.\n\nThis endpoint uses a Langchain pipeline with OpenAI GPT models to generate\ncontextually-aware layer definitions based on organizational context, hierarchical\nrelationships, contained domains, and layer purpose information.\nSupports optional flavor parameter to use different LLM configurations.",
        "operationId": "suggest_layer_definition_api_llm_suggest_layer_definition_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayerDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayerLLMSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/llm/suggest_domain_definition": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Suggest Domain Definition",
        "description": "Generate a domain definition suggestion based on provided context using LLM.\n\nThis endpoint uses a Langchain pipeline with OpenAI GPT models to generate\ncontextually-aware domain definitions based on thematic scope, hierarchical\nrelationships, contained terms, and domain boundaries information.\nSupports optional flavor parameter to use different LLM configurations.",
        "operationId": "suggest_domain_definition_api_llm_suggest_domain_definition_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainLLMSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DefinitionSuggestionRequest": {
        "properties": {
          "term": {
            "type": "string",
            "title": "Term",
            "description": "Term to be analyzed"
          },
          "domain_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Title",
            "description": "Title of the domain"
          },
          "domain_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Definition",
            "description": "Definition of the domain"
          },
          "parent_term_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Term Title",
            "description": "Title of the parent term"
          },
          "parent_term_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Term Definition",
            "description": "Definition of the parent term"
          },
          "parent_relationship_predicate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Relationship Predicate",
            "description": "Relationship predicate to parent"
          },
          "component_terms": {
            "items": {
              "$ref": "#/components/schemas/ComponentTerm"
            },
            "type": "array",
            "title": "Component Terms",
            "description": "Component terms with their context"
          },
          "current definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Definition",
            "description": "Current definition if any"
          },
          "dbpedia_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Dbpedia Context",
            "description": "DBpedia context information"
          },
          "wikidata_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Wikidata Context",
            "description": "Wikidata context information"
          },
          "flavor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flavor",
            "description": "Flavor ID or title to use for this request"
          }
        },
        "type": "object",
        "required": [
          "term"
        ],
        "title": "DefinitionSuggestionRequest",
        "description": "Request model for term definition suggestion endpoint"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMSuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Always true for success responses",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/DefinitionSuggestionResponse",
            "description": "The response data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "LLMSuccessResponse",
        "description": "Success response wrapper for LLM endpoints"
      },
...
      "LayerDefinitionRequest": {
        "properties": {
          "layer_title": {
            "type": "string",
            "title": "Layer Title",
            "description": "Title of the layer to be analyzed"
          },
          "layer_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer Description",
            "description": "Current description of the layer"
          },
          "parent_layer_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Layer Title",
            "description": "Title of the parent layer"
          },
          "parent_layer_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Layer Definition",
            "description": "Definition of the parent layer"
          },
          "contained_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contained Domains",
            "description": "List of domain titles contained in this layer"
          },
          "layer_purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer Purpose",
            "description": "Purpose or role of this layer"
          },
          "current_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Definition",
            "description": "Current definition if any"
          },
          "reference_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Reference Context",
            "description": "Additional reference context"
          },
          "flavor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flavor",
            "description": "Flavor ID or title to use for this request"
          }
        },
        "type": "object",
        "required": [
          "layer_title"
        ],
        "title": "LayerDefinitionRequest",
        "description": "Request model for layer definition suggestion endpoint"
      },
...
      "LayerLLMSuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Always true for success responses",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/LayerDefinitionResponse",
            "description": "The response data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "LayerLLMSuccessResponse",
        "description": "Success response wrapper for layer LLM endpoints"
      },
...
    }
  }
}